pyproject.toml
[project]
name = "drf-shop"
version = "0.1.0"
description = ""
authors = [
    {name = "dk"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
